+++ imports/api/migrations/index.js
@@ -1,5 +1,6 @@
 import moment from 'moment';
 import tools from '/imports/startup/tools/index';
+import schema from '/imports/startup/schema/index';
 
 import { Accessories } from '/imports/api/accessories/index';
 import { Clients } from '/imports/api/clients/index';
@@ -15,8 +16,36 @@ import { Series } from '/imports/api/series/index';
 
 if (Meteor.isServer) {
   Migrations.add({
+    version: 2,
+    name: "Adapt Users to SimpleSchema",
+    up: function() {
+      var users = Meteor.users.find({}).fetch();
+      users.forEach((user) => {
+        var _id = user._id;
+        var newUser = {
+          ...user,
+          profile: {
+            firstName: user.firstName,
+            lastName: user.lastName,
+            type: user.type
+          }
+        }
+        delete newUser.firstName;
+        delete newUser.lastName;
+        delete newUser.type;
+        delete newUser.pages;
+        delete newUser._id;
+
+        schema('users', 'insert').validate(newUser);
+
+        Meteor.users.update({ _id }, newUser);
+      })
+    },
+    down: function() {}
+  });
+  Migrations.add({
     version: 1,
-    name: "Constrains Accessories to SimpleSchema",
+    name: "Adapt Accessories to SimpleSchema",
     up: function() {
       Accessories.update(
         {},