+++ imports/api/users/index.js
@@ -1,11 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 import { Accounts } from 'meteor/accounts-base';
-import SimpleSchema from 'simpl-schema';
+import schema from '/imports/startup/schema/index';
 import tools from '/imports/startup/tools/index';
-
-var passwordMinLength = 4;
-var userNameMinLength = 3;
-var userNameMaxLength = 40;
 
 Meteor.users.deny({
   insert() { return true; },
@@ -16,94 +12,77 @@ Meteor.users.deny({
 if (Meteor.isServer) {
   Meteor.publish('usersPub', function () {
     if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
-    return Meteor.users.find(
-      { visible: true },
-      { fields: { type: 1, emails: 1, firstName: 1, lastName: 1 } },
-      { sort: { username: 1 } }
-    );
+    return Meteor.users.find({});
   });
 
   Meteor.methods({
     'users.insert'(state) {
       if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
-      var firstName = state.firstName;
-      var lastName = state.lastName;
-      var username = state.username;
-      var password = state.password;
-      var type = state.type;
-      var emails = [{address: state.emails, verified: false}];
-
-      if (!username || !emails) {
-        throw new Meteor.Error('required-fields-empty');
-      };
-      if (password.length < passwordMinLength) {
-        throw new Meteor.Error('password-too-short');
-      };
-      if (username.length < userNameMinLength) {
-        throw new Meteor.Error('name-too-short');
-      };
-      if (username.length > userNameMaxLength) {
-        throw new Meteor.Error('name-too-long');
-      };
-      const _id = Accounts.createUser({username, emails, password});
-      if (!_id) {throw new Meteor.Error('user-not-created');}
-      const data = {
-        _id,
-        firstName,
-        lastName,
-        type,
-        username,
-        emails,
-        visible: true
-      };
-      Meteor.users.update({ _id }, { $set: data });
-      Meteor.call('history.insert', data, 'users');
-    },
-
-    'users.hide'(_id) {
-      if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
-      const data = {
-        _id,
-        visible: false
+      var data = {
+        username: state.username,
+        password: state.password,
+        email: state.email,
+        profile: {
+          firstName: state.profile.firstName,
+          lastName: state.profile.lastName,
+          type: state.profile.type
+        }
       }
-      Meteor.users.update({ _id }, { $set: data });
-      Meteor.call('history.insert', data, 'users');
+      schema('users', 'update').validate(data);
+      var _id = Accounts.createUser(data);
+      if (!_id) {throw new Meteor.Error('user-not-created')}
+      delete data.password;
+      Meteor.call('history.insert', {...data, _id}, 'users.insert');
     },
-
+    // 'users.hide'(_id) {
+    //   if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
+    //   const data = {
+    //     _id,
+    //     visible: false
+    //   }
+    //   Meteor.users.update({ _id }, { $set: data });
+    //   Meteor.call('history.insert', data, 'users');
+    // },
     'users.update'(state) {
       if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
-      var _id = state._id;
-      var firstName = state.firstName;
-      var lastName = state.lastName;
-      var username = state.username;
-      var password = state.password;
-      var type = state.type;
-      var emails = [{address: state.emails, verified: false}];
-
-      if (!username || !emails) {
-        throw new Meteor.Error('required-fields-empty');
-      };
-      if (password && password.length < passwordMinLength) {
-        throw new Meteor.Error('password-too-short');
-      };
-      if (username.length < userNameMinLength) {
-        throw new Meteor.Error('name-too-short');
-      };
-      if (username.length > userNameMaxLength) {
-        throw new Meteor.Error('name-too-long');
-      };
-      const data = {
-        firstName,
-        lastName,
-        type,
-        username,
-        emails,
-        visible: true
+      if (!Meteor.user().profile.type !== 'administrator') {
+        throw new Meteor.Error('unauthorized');
       }
+      var _id = state._id;
+      var data = {
+        username: state.username,
+        email: state.email,
+        profile: {
+          firstName: state.profile.firstName,
+          lastName: state.profile.lastName,
+          type: state.profile.type
+        }
+      }
+      schema('users', 'update').validate(data);
+
       Meteor.users.update({ _id }, { $set: data });
       data._id = _id;
       Meteor.call('history.insert', data, 'users');
-      password ? Accounts.setPassword(_id, password) : null;
+      return true;
+    },
+    'users.setPassword'(oldPassword, newPassword, userId) {
+      if (!Meteor.userId()) throw new Meteor.Error('unauthorized');
+      if (userId !== Meteor.userId()) {
+        throw new Meteor.Error('unauthorized');
+      }
+      schema('users', 'password').validate({
+        password: newPassword
+      });
+      var res = Accounts._checkPassword(Meteor.user(), oldPassword)
+      if (res.error) {
+        console.log(res);
+        throw res.error;
+      } else {
+        Accounts.setPassword(Meteor.userId(), newPassword, {
+          logout: false
+        });
+        return true;
+      }
     }
   })
 }